Configurando a Exception na Service

+++ pwnned/src/main/java/com/pwnned/domain/service/UserService.java
@@ -1,5 +1,6 @@
 package com.pwnned.domain.service;
 
+import com.pwnned.domain.exception.UserNotFoundException;
 import com.pwnned.domain.model.User;
 import com.pwnned.port.input.UserServicePort;
 import com.pwnned.port.output.UserRepositoryPort;
@@ -7,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 @Service
 public class UserService implements UserServicePort {
@@ -24,21 +26,29 @@ public class UserService implements UserServicePort {
 
     @Override
     public List<User> getAllUsers() {
-        return userRepositoryPort.findAll();
+        List<User> users = userRepositoryPort.findAll();
+        if (users.isEmpty()) throw new UserNotFoundException("Users Not Found");
+        return users;
     }
 
     @Override
-    public Optional<User> getSingleUser(Long userId) {
-        return userRepositoryPort.findById(userId);
+    public Optional<User> getSingleUser(UUID userId) {
+        Optional<User> user = userRepositoryPort.findById(userId);
+        if (user.isEmpty()) throw new UserNotFoundException("User " + userId + " Not Found");
+        return user;
     }
 
     @Override
-    public void deleteUser(Long userId) {
+    public void deleteUser(UUID userId) {
+        Optional<User> user = userRepositoryPort.findById(userId);
+        if (user.isEmpty()) throw new UserNotFoundException("User " + userId + " Not Found");
         userRepositoryPort.deleteById(userId);
     }
 
     @Override
     public void deleteAllUsers() {
+        List<User> users = userRepositoryPort.findAll();
+        if (users.isEmpty()) throw new UserNotFoundException("No Users to Delete");
         userRepositoryPort.deleteAll();
     }
 }
